edit: 회원가입 서버 통신 기능 수정
