+++ src/pages/signup/signUp.html
@@ -5,8 +5,8 @@
     <meta charset="UTF-8" />
     <link rel="icon" type="image/svg+xml" href="/taing.ico" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <script type="module" src="/main.js"></script>
-    <script type="module" src="/src/js/signUp.js"></script>
+    <link rel="stylesheet" href="/src/styles/style.scss" />
+    <script type="module" src="/src/pages/signup/signUp.js"></script>
   </head>
 
   <body>
